backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Rate limiting
slowapi==0.1.9

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2

# For JWT token generation for skills
PyJWT==2.8.0

# JSON handling
orjson==3.9.10